+++ Backend/database/migrations/2025_01_16_090735_add_competition_id_column_to_mentors_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
         }
 
             Schema::table('mentors', function (Blueprint $table) {
-            $table->foreignId('competition_id')->constrained()->onDelete('cascade');
+            if (!Schema::hasColumn('mentors', 'competition_id')) { $table->foreignId('competition_id')->constrained()->onDelete('cascade'); }
         });
         }
     }
