+++ Backend/database/migrations/2025_01_19_170512_add_application_id_column_to_teams_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
         }
 
             Schema::table('teams', function (Blueprint $table) {
-            $table->foreignId('application_id')
+            if (!Schema::hasColumn('teams', 'application_id')) { $table->foreignId('application_id') }
                 
                 ->constrained('competition_applications')
                 ->onDelete('cascade');
