+++ Backend/database/migrations/2025_07_30_142628_add_stage_id_to_project_evaluations_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('project_evaluations')) {
             Schema::table('project_evaluations', function (Blueprint $table) {
-            $table->foreignId('stage_id')
+            if (!Schema::hasColumn('project_evaluations', 'stage_id')) { $table->foreignId('stage_id') }
                 ->constrained()
                 ->onDelete('cascade');
         });
